+++ app/Fixers/LaravelBlade/Fixer.php
@@ -44,15 +44,6 @@ class Fixer extends AbstractFixer implements HasPrettierDependencies
     /**
      * {@inheritdoc}
      */
-    public function finderNames(): array
-    {
-        // Blade files are ".php" files, which the finder already discovers.
-        return [];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getName(): string
     {
         return 'Pint/laravel_blade';